requirement.txt
# Use bleeding edge instead of stable
# tweety-ns
git+https://github.com/mahrtayyab/tweety.git@main

# does not work with Free Twitter acc
# tweepy

pyyaml
pandas
kafka-python
pyspark
python-dotenv
apache-airflow
virtualenv
elasticsearch
elasticsearch-hadoop